Now able to scroll with the keyboard.
I like some of the shortcuts to be alt and some to be ctrl.  Don't need
the change tab shortcuts (use others that are global for all
applications).  Hardcoded some of the shortcuts too - I don't need the
preferences and I got tired of setting them after reloading the
extension.